+++ src/client/utils/tracking/hooks/useTrackOfferEvent.ts
@@ -1,0 +1,70 @@
+import { useCallback } from 'react'
+import { QuoteCartDocument, QuoteCartQuery } from 'data/graphql'
+import { useQuoteCartIdFromUrl } from 'utils/hooks/useQuoteCartIdFromUrl'
+import { useSelectedInsuranceTypes } from 'utils/hooks/useSelectedInsuranceTypes'
+import {
+  getMonthlyCostDeductionIncentive,
+  getSelectedBundleVariant,
+} from 'api/quoteCartQuerySelectors'
+import { useCurrentLocale } from 'l10n/useCurrentLocale'
+import { apolloClient } from 'apolloClient'
+import {
+  EventParameters,
+  trackOfferEvent,
+} from 'utils/tracking/trackOfferEvent'
+
+export const useTrackOfferEvent = () => {
+  const { isoLocale } = useCurrentLocale()
+
+  const { quoteCartId } = useQuoteCartIdFromUrl()
+
+  const [selectedInsuranceTypes] = useSelectedInsuranceTypes()
+
+  const trackEventHandler = useCallback(
+    ({ eventName, options = {} }: EventParameters) => {
+      const runQuery = async () => {
+        if (apolloClient && quoteCartId) {
+          try {
+            const quoteCartQuery = await apolloClient!.client.query<
+              QuoteCartQuery
+            >({
+              query: QuoteCartDocument,
+              variables: {
+                id: quoteCartId,
+                locale: isoLocale,
+              },
+            })
+            return quoteCartQuery.data
+          } catch (e) {
+            return
+          }
+        } else return
+      }
+      const trackEventCallback = async () => {
+        const quoteCartQueryData = await runQuery()
+        const isReferralCodeUsed =
+          getMonthlyCostDeductionIncentive(quoteCartQueryData) !== undefined
+        const selectedBundleVariant = getSelectedBundleVariant(
+          quoteCartQueryData,
+          selectedInsuranceTypes,
+        )
+
+        if (selectedBundleVariant) {
+          trackOfferEvent(
+            eventName,
+            selectedBundleVariant.bundle,
+            isReferralCodeUsed,
+            {
+              quoteCartId,
+              ...options,
+            },
+          )
+        }
+      }
+      trackEventCallback()
+    },
+    [quoteCartId, isoLocale, selectedInsuranceTypes],
+  )
+
+  return trackEventHandler
+}
